refactor: add centralized style system
