+++ src/main.js
@@ -1,4 +1,5 @@
 import { createApp } from 'vue'
+import './styles/variables.css'
 import './style.css'
 import App from './App.vue'
 import { pinia } from './store'
